+++ python/rt2d/path_family/__init__.py
@@ -1,11 +1,12 @@
 from .api import compute_rx_partition, compute_rx_partition_runtime, compute_rx_rays_runtime
 from .runtime import PathFamilyRuntime, build_path_family_runtime
-from .types import PathFamily, RayHit, ReflectionInteractionRef
+from .types import DiffractionInteractionRef, PathFamily, RayHit, ReflectionInteractionRef
 
 __all__ = [
     "PathFamily",
     "PathFamilyRuntime",
     "RayHit",
+    "DiffractionInteractionRef",
     "ReflectionInteractionRef",
     "build_path_family_runtime",
     "compute_rx_partition",